requirements.txt
passlib
django
python-ldap
django-bootstrap4
django-auth-ldap
django-guardian
django-markdownify
django-taggit
django-impersonate
psycopg2
